Enable builds with -std=c99 and -pedantic.
Originally -ansi flag was intended but it turned out that snprintf is not
supported in C90 standard, which is a show stopper. Therefore we settled for
-std=c99 and -pendantic, with the notable exception of src/libsnoopy.c.

Exception:
It was pointed out that dlsym returns object pointer, which may not be
cast into function pointer. Therefore only for this file (libsnoopy.c) we
keep -pendantic disabled.

Additionally some minor adjustments had to be done to make build complete
flawlessly.