tests: added combined tests, which test Snoopy as a whole
These tests load alternative libsnoopy-test.so, which parses environment
variabl SNOOPY_INI and loads config file from that location, provided that
it exists and it is readable. From there on, it acts as normal Snoopy.

Generally this test suite uses file output to check what actually comes out
of Snoopy.