Refactoring: make arguments to datasources, filters and outputs constant where applicable
Motivation:
When calling filter exclude_spawns_of manually with constant string as filterArg,
there was a segfault because this filter relied on modifying filterArg when it
parsed it.

Reason:
Depending on the implementation of option parser, this might become problematic.

Therefore all strings that datasources, filters and outputs do not need to modify
have been marked as constant strings. The only ones left available to modify are:
- datasource(result): modifiable content, but not pointer itself;
- filter(logMessage): same as above.