Refactoring init: now explicit methods for enabling/disabling config file parsing are available
This is a runtime method, not configuration flag. This is needed
for tests which must call snoopy_init() now (since thread-safety
was introduced). Init() in turn tries to parse config file on the
final location. If that file is missing (on Travis CI VMs for
example), some tests fail (i.e. stderr).

Now the so-called preinit methods were introduced, to be used before
snoopy_init() is called.