New datasource: snoopy_threads
Returns number of threads within a single program, that Snoopy is currently
configured for.

This datasource will return "1" in most of cases. The only time it might
return more than 1 is when two threads try to execute external program at once.

This seems imposible, as if program forks, all data is cloned into separate
process. Within that process the execv() is called and Snoopy is triggered in
this new process. This means that it should not interfere with Snoopy data in
the original process, even with the global variables.

Well, unless global variables in shared libraries are shared across all
processes, but this is not the default Linux behaviour.