TSRM: when counting threads, move check for uninitialized repo inside mutex block
Reason:
One thread might uninitialize repo in between and then accessing
->count member causes segmentation fault.

THIS IS A KNOWINGLY BROKEN COMMIT!

Moving check into mutex block would be a normal thing to do, but doing so
causes a deadlock for which I can not figure out the source/condition ATM.

Next commit will fix this issue with a workaround, which was stumbled upon
accidentally and actually I do not know why it works. Will look into it.