Threading: add new test to repeatedly check for segmentation fault. Not included in test suite by default.
Reason for creating:
There were some threading issues with Snoopy (resolved in previous commit) which
required multiple invocations to be triggered. It occured on every 30th invocation
on average (number determined manually, using "feeling" and "memory" meter).

Reasons for not including it in standard test suite:
It takes significant amount of time (10-30 seconds) compared to other
tests (0.1-1 second). If threading issue is suspected, it might be
run manually to try to trigger and confirm it.