Test/threads/segfault: reenable test in test suite, faster; test now uses mutex for counting threads.
All output that was going on in loops was disabled, so threading implementation
is tested under higher pressure - less output handling, faster it has to be doing
the thread management tasks (mutexes, etc).

Test runs now in under a second so there is no reason to not enable it in general
test suite.

Also, bumped threads to 200 in test.