lib/liblc/.../list.c: fix bug when pointers to list entries that were removed were not set to NULL
Finally, this was the reason for all deadlocks in thread-safe Snoopy
implementation. When you look at the code, if you remove node from either
list start or end, the corresponding pointers at first (for prev node) or last
(for next node) node were not set to null. This caused foreach to go into
infinite loop.
The list implementation is shit, if I judge it by this error. But, as it is
FOSS, I guess we got what we paid for :)