snoopy-install.sh file for off-line use.
I administer many systems, several of which by design & requirement are NOT connected to the internet and have no way of communicating with the internet. This causes several big issues with installing and/or updating snoopy-logger.

To resolve this issue, I modified the snoopy-installer.sh script to add two new install "modes". 
The first is run by: `snoopy-installer.sh download`. This, when run from an internet-connected system, will download the latest stable snoopy-*.tar.gz file to the same directory from which the script is run.

The second is run by: `snoopy-installer.sh local`. This, when copied to the non-internet-connected machine _with_ the previously downloaded snoopy-*.tar.gz file, will build and install the latest stable snoopy release from the gzipped-tar file.

**Note:** I have ONLY validated the functionality of the script for `download`, `local`, `stable` invocations. I use this new script on my several systems. It takes a lot of frustration and work out of the mix when installing/upgrading snoopy.

Updated script code is attached as: `snoopy-install-NEW.sh`. 
(I'm new to github and this project, so I apologize in advance if I'm submitting this incorrectly.)