GH #145: Enable analysis of PID 1 in filter `exclude_spawns_of`
The implementation of exclude_spawns_of somewhat predates widespread usage of
Linux containers. Before containers, PID 1 was always some sort of an init
process, always there, always called `init`, and it somewhat made sense to
just ignore it.

But with the onset of containers/namespaces, the PID 1 inside an i.e. Docker
container can be any executable that is configured as container's entrypoint.

The solution is to adjust the filter's method and it perform the process name
analysis all the way up to and including PID 1, whatever that process may be.