core/ds/domain: Fix incorrect handling of comments in /etc/hosts
According to /etc/hosts' manpage, comments that don't start at the beginning of
a line should be allowed, but the `hostname` command (at least on Ubuntu 20.04),
or more likely the underlying `getaddrinfo()` function, doesn't actually return
the domain name in that case. Regardless, let's do the right thing here and
extract the domain name successfully anyway.

(We might want to switch to using `getaddrinfo()` in the future, instead of
"manually" parsing the `/etc/hosts` file.)