datasource/cmdline: Fix incorrect max string length determination
This code has been tampered with in the following commits, all related to
fixes of SonarCloud-reported code smells:
- f73ea2a11f6d0c3a73a8d643ca7732d86ac2fff9
- 0cdb9605f32d62499d6609ae866fab54f28a3b56
- fb1d38859e8ef21c1f039d9bd83aeddb7fea8252
- 316541fa6601b40ec4d369578aab2f8dda123ca2

These changes resulted in an incorrect long->int conversion _before_ the
value comparison. Earlier, when the conversion was implicit, the compiler
did the right thing and compared the values using the long type instead.