ci: Error out if an unknown `./configure` option is encountered
GNU suggests this option should not be enabled by default, since identical
options are passed into subdirectories containing `configure` scripts (i.e.
if someone is importing Snoopy into their main project and building it in
the main project's subdirectory).

(That said, I wanted to enable this option for everybody regardless of the
suggestion above, but I could not find a way to do it.)