doc/releasing: Add instruction to check the package install guide before releasing
Post-release, binary packages are built and then in the last step, installation
of those packages is tested. To test the process out, shell code is extracted from
the relevant document and applied to individual distributions to verify if Snoopy
installs correctly with it.

Therefore, we need to have a matching set of supported distributions in Actions
configuration and in this guide document.